agent/pyproject.toml
[project]
name = "openclaw-audit-agent"
version = "0.1.0"
description = "Multi-agent AI maturity audit backend (Claude Agent SDK)"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "anthropic>=0.40.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "sse-starlette>=2.1.0",
    "pydantic>=2.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
